refactor: Controller 내부에서 HttpRequest의 getSession() 메소드를 호출하여 HttpSession을 가져오도록 구현
