+++ app/src/main/java/nextstep/jwp/controller/IndexController.java
@@ -1,11 +1,11 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.request.HttpRequest;
-import nextstep.jwp.web.model.HttpSession;
+import nextstep.jwp.response.HttpResponse;
 
 public class IndexController implements Controller {
 
-    public String index(HttpRequest httpRequest, HttpSession httpSession) {
+    public String index(HttpRequest httpRequest, HttpResponse httpResponse) {
         return "index.html";
     }
 }
