+++ app/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -3,13 +3,15 @@ package nextstep.jwp.controller;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import nextstep.jwp.request.HttpRequest;
+import nextstep.jwp.response.HttpResponse;
 import nextstep.jwp.web.model.HttpSession;
 
 import java.util.Objects;
 
 public class RegisterController implements Controller {
 
-    public String registerPage(HttpRequest httpRequest, HttpSession httpSession) {
+    public String registerPage(HttpRequest httpRequest, HttpResponse httpResponse) {
+        HttpSession httpSession = httpRequest.getSession(httpResponse);
         User user = getUser(httpSession);
         if (Objects.nonNull(user)) {
             return "redirect: /index.html";
@@ -17,7 +19,7 @@ public class RegisterController implements Controller {
         return "register.html";
     }
 
-    public String register(HttpRequest httpRequest, HttpSession httpSession) {
+    public String register(HttpRequest httpRequest, HttpResponse httpResponse) {
         String account = httpRequest.getAttribute("account");
         String password = httpRequest.getAttribute("password");
         String email = httpRequest.getAttribute("email");
