+++ app/src/main/java/nextstep/jwp/request/HttpRequest.java
@@ -1,6 +1,8 @@
 package nextstep.jwp.request;
 
+import nextstep.jwp.response.HttpResponse;
 import nextstep.jwp.web.model.HttpCookie;
+import nextstep.jwp.web.model.HttpSession;
 
 public interface HttpRequest {
 
@@ -9,4 +11,5 @@ public interface HttpRequest {
     String getResourceName();
     String getAttribute(String name);
     HttpCookie getCookies();
+    HttpSession getSession(HttpResponse response);
 }
