+++ app/src/main/java/nextstep/jwp/web/FrontController.java
@@ -1,15 +1,10 @@
 package nextstep.jwp.web;
 
 import nextstep.jwp.exception.InternalServerError;
-import nextstep.jwp.exception.InvalidHttpSessionException;
 import nextstep.jwp.exception.PageNotFoundError;
 import nextstep.jwp.request.HttpRequest;
 import nextstep.jwp.response.HttpResponse;
 import nextstep.jwp.response.HttpStatusCode;
-import nextstep.jwp.web.model.Cookie;
-import nextstep.jwp.web.model.HttpCookie;
-import nextstep.jwp.web.model.HttpSession;
-import nextstep.jwp.web.model.HttpSessions;
 
 import java.util.Optional;
 
@@ -28,8 +23,7 @@ public class FrontController {
             Optional<ControllerMethod> optionalControllerMethod = requestMapping.getControllerMethod(httpRequest);
             if (optionalControllerMethod.isPresent()) {
                 ControllerMethod controllerMethod = optionalControllerMethod.orElseThrow(PageNotFoundError::new);
-                HttpSession httpSession = this.getHttpSession(httpRequest, httpResponse);
-                String viewName = (String) controllerMethod.invoke(httpRequest, httpSession);
+                String viewName = (String) controllerMethod.invoke(httpRequest, httpResponse);
                 httpResponse.setView(viewName, HttpStatusCode.OK);
                 return;
             }
@@ -40,25 +34,4 @@ public class FrontController {
             httpResponse.setView(NOT_FOUND_ERROR_PAGE, HttpStatusCode.NOTFOUND);
         }
     }
-
-    private HttpSession getHttpSession(HttpRequest httpRequest, HttpResponse httpResponse) {
-        HttpCookie httpCookie = httpRequest.getCookies();
-        if (hasValidJSessionId(httpCookie)) {
-            return HttpSessions.getSession(httpCookie.getJSessionId())
-                    .orElseThrow(() -> new InvalidHttpSessionException("세션이 존재하지 않습니다."));
-        }
-
-        HttpSession httpSession = HttpSession.create();
-        httpResponse.addCookie(new Cookie("JSESSIONID", httpSession.getId()));
-        return HttpSessions.save(httpSession);
-    }
-
-    private boolean hasValidJSessionId(HttpCookie httpCookie) {
-        if (httpCookie.hasJSessionId()) {
-            String jSessionId = httpCookie.getJSessionId();
-            return HttpSessions.isExistsId(jSessionId);
-        }
-        return false;
-    }
-
 }
